olorin-media/bayit-plus/backend/pyproject.toml
[project]
name = "bayit-plus-backend"
version = "0.1.0"
description = "Bayit+ Hebrew streaming platform API"
authors = [
    {name = "Bayit+",email = "admin@bayit.tv"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "motor (>=3.7.1,<4.0.0)",
    "pymongo (>=4.15.5,<5.0.0)",
    "beanie (>=2.0.1,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.1.0)",
    "bcrypt (>=4.0.0,<5.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "stripe (>=14.1.0,<15.0.0)",
    "anthropic (>=0.75.0,<0.76.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiohttp (>=3.13.3,<4.0.0)",
    "beautifulsoup4 (>=4.12.0,<5.0.0)",
    "redis (>=5.0.0,<6.0.0)",
    "apscheduler (>=3.10.0,<4.0.0)",
    "pytz (>=2024.1)",
    "webvtt-py (>=0.4.0,<0.5.0)",
    "livekit-api (>=0.4.0,<0.5.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "feedparser (>=6.0.12,<7.0.0)",
    "google-cloud-storage (>=2.10.0,<3.0.0)",
    "google-auth (>=2.23.0,<3.0.0)",
    "google-cloud-speech (==2.21.0)",
    "google-cloud-translate (==3.12.0)",
    "ffmpeg-python (==0.2.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "openai (>=2.15.0,<3.0.0)",
    "twilio (>=9.9.1,<10.0.0)",
    "phonenumbers (>=9.0.21,<10.0.0)",
    "xmltodict (>=0.14.2,<0.15.0)",
    "python-dateutil (>=2.9.0,<3.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "python-chess (>=1.10.0,<2.0.0)",
    "chess (>=1.11.2,<2.0.0)",
    "websockets (>=15.0.0,<16.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "sentry-sdk[fastapi] (>=1.40.0,<2.0.0)",
    "webauthn (>=2.2.0,<3.0.0)",
    "pinecone-client (>=3.0.0,<4.0.0)",
    "bayit-voice-pipeline @ file:///Users/olorin/Documents/Bayit-Plus/packages/bayit-voice-pipeline",
    "bayit-translation @ file:///Users/olorin/Documents/Bayit-Plus/packages/bayit-translation",
    "olorin-core @ file:///Users/olorin/Documents/Bayit-Plus/packages/olorin-core",
    "olorin-shared @ file:///Users/olorin/Documents/Bayit-Plus/packages/olorin-shared",
    "olorin-i18n @ file:///Users/olorin/Documents/Bayit-Plus/packages/python/olorin-i18n"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=4.1.0,<5.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "black (>=24.0.0,<25.0.0)",
    "isort (>=5.13.0,<6.0.0)",
    "pylint (>=3.0.0,<4.0.0)",
    "pip-audit (>=2.7.0,<3.0.0)"
]
